+++ api/src/opentrons/protocol_api/core/engine/instrument.py
@@ -6,7 +6,7 @@ from typing import Optional, TYPE_CHECKING
 from opentrons.types import Location, Mount
 from opentrons.hardware_control import SyncHardwareAPI
 from opentrons.hardware_control.dev_types import PipetteDict
-from opentrons.protocols.api_support.util import Clearances, PlungerSpeeds, FlowRates
+from opentrons.protocols.api_support.util import PlungerSpeeds, FlowRates
 from opentrons.protocol_engine import DeckPoint, WellLocation
 from opentrons.protocol_engine.clients import SyncClient as EngineClient
 from opentrons.protocols.api_support.definitions import MAX_SUPPORTED_VERSION
@@ -40,10 +40,6 @@ class InstrumentCore(AbstractInstrument[WellCore]):
         self._sync_hardware_api = sync_hardware_api
         self._protocol_core = protocol_core
 
-        # TODO(jbl 2022-11-09) clearances should be move out of the core
-        self._well_bottom_clearances = Clearances(
-            default_aspirate=1.0, default_dispense=1.0
-        )
         # TODO(jbl 2022-11-03) flow_rates should not live in the cores, and should be moved to the protocol context
         #   along with other rate related refactors (for the hardware API)
         self._flow_rates = FlowRates(self)
@@ -363,7 +359,7 @@ class InstrumentCore(AbstractInstrument[WellCore]):
         return self._sync_hardware_api.get_attached_instrument(self.get_mount())  # type: ignore[no-any-return]
 
     def get_channels(self) -> int:
-        return self._engine_client.state.pipettes.get_channels(self._pipette_id)
+        return self._engine_client.state.tips.get_pipette_channels(self._pipette_id)
 
     def has_tip(self) -> bool:
         return self.get_hardware_state()["has_tip"]
@@ -376,9 +372,6 @@ class InstrumentCore(AbstractInstrument[WellCore]):
 
     def get_return_height(self) -> float:
         raise NotImplementedError("InstrumentCore.get_return_height not implemented")
-
-    def get_well_bottom_clearance(self) -> Clearances:
-        return self._well_bottom_clearances
 
     def get_speed(self) -> PlungerSpeeds:
         raise NotImplementedError("InstrumentCore.get_speed not implemented")