+++ api/src/opentrons/protocol_api/labware.py
@@ -18,7 +18,6 @@ from opentrons.types import Location, Point, LocationLabware
 from opentrons.protocols.api_support.types import APIVersion
 from opentrons.protocols.api_support.util import requires_version
 from opentrons.protocols.api_support.definitions import MAX_SUPPORTED_VERSION
-from opentrons.protocols.geometry.deck_item import DeckItem
 from opentrons.protocols.geometry.well_geometry import WellGeometry
 
 # TODO(mc, 2022-09-02): re-exports provided for backwards compatibility
@@ -96,10 +95,16 @@ class Well:
     @property  # type: ignore[misc]
     @requires_version(2, 0)
     def has_tip(self) -> bool:
+        """If parent labware is a tip rack, whether this well contains a tip."""
         return self._impl.has_tip()
 
     @has_tip.setter
     def has_tip(self, value: bool) -> None:
+        _log.warning(
+            "Setting the `Well.has_tip` property manually has been deprecated"
+            " and will raise an error in a future version of the Python Protocol API."
+        )
+
         self._impl.set_has_tip(value)
 
     @property
@@ -230,7 +235,7 @@ class Well:
         return hash(self.top().point)
 
 
-class Labware(DeckItem):
+class Labware:
     """
     This class represents a labware, such as a PCR plate, a tube rack,
     reservoir, tip rack, etc. It defines the physical geometry of the labware,
@@ -292,7 +297,11 @@ class Labware(DeckItem):
 
     @property
     def separate_calibration(self) -> bool:
-        return self._implementation.separate_calibration
+        _log.warning(
+            "Labware.separate_calibrations is a deprecated internal property."
+            " It no longer has meaning, but will always return `False`"
+        )
+        return False
 
     @property  # type: ignore
     @requires_version(2, 0)
@@ -324,6 +333,7 @@ class Labware(DeckItem):
         load it, or the label of the labware specified by a user."""
         return self._implementation.get_name()
 
+    # TODO(jbl, 2022-12-06): deprecate officially when there is a PAPI version for the engine core
     @name.setter
     def name(self, new_name: str) -> None:
         """Set the labware name"""
@@ -613,6 +623,7 @@ class Labware(DeckItem):
     def tip_length(self) -> float:
         return self._implementation.get_tip_length()
 
+    # TODO(jbl, 2022-12-06): deprecate officially when there is a PAPI version for the engine core
     @tip_length.setter
     def tip_length(self, length: float) -> None:
         self._implementation.set_tip_length(length)