+++ api/src/opentrons/protocol_engine/commands/calibration/__init__.py
@@ -1,4 +1,11 @@
 """Calibration commands."""
+from .calibrate_gripper import (
+    CalibrateGripperParams,
+    CalibrateGripperResult,
+    CalibrateGripperCreate,
+    CalibrateGripperCommandType,
+    CalibrateGripper,
+)
 
 from .calibrate_pipette import (
     CalibratePipetteParams,
@@ -17,6 +24,13 @@ from .move_to_maintenance_position import (
 )
 
 __all__ = [
+    # calibration/calibrateGripper
+    "CalibrateGripper",
+    "CalibrateGripperCreate",
+    "CalibrateGripperParams",
+    "CalibrateGripperParamsProbe",
+    "CalibrateGripperResult",
+    "CalibrateGripperCommandType",
     # calibration/calibratePipette
     "CalibratePipette",
     "CalibratePipetteCreate",