+++ app/src/organisms/ProtocolDetails/__tests__/utils.test.ts
@@ -1,0 +1,24 @@
+import {
+  TC_MODULE_LOCATION_OT2,
+  TC_MODULE_LOCATION_OT3,
+  OT2_STANDARD_MODEL,
+  OT3_STANDARD_MODEL,
+} from '@opentrons/shared-data'
+import { getSlotsForThermocycler } from '../utils'
+
+describe('getSlotsForThermocylcer', () => {
+  it('returns the correct thermocylcer slots given an OT3', () => {
+    const result = getSlotsForThermocycler(OT3_STANDARD_MODEL)
+    expect(result).toEqual(TC_MODULE_LOCATION_OT3)
+  })
+
+  it('returns the correct thermocylcer slots given an OT2', () => {
+    const result = getSlotsForThermocycler(OT2_STANDARD_MODEL)
+    expect(result).toEqual(TC_MODULE_LOCATION_OT2)
+  })
+
+  it('returns OT2 thermocycler location when robotType is not available', () => {
+    const result = getSlotsForThermocycler(null)
+    expect(result).toEqual(TC_MODULE_LOCATION_OT2)
+  })
+})