+++ app/src/pages/OnDeviceDisplay/__tests__/ConnectedNetworkInfo.test.tsx
@@ -31,7 +31,7 @@ jest.mock('react-router-dom', () => {
 const render = (path = '/') => {
   return renderWithProviders(
     <MemoryRouter initialEntries={[path]} initialIndex={0}>
-      <Route path="/connected-network-info/:ssid">
+      <Route path="/network-setup/wifi/connected-network-info/:ssid">
         <ConnectedNetworkInfo />
       </Route>
     </MemoryRouter>,
@@ -48,13 +48,6 @@ const initialMockWifi = {
   type: Networking.INTERFACE_WIFI,
 }
 
-const initialMockEthernet = {
-  ipAddress: '127.0.0.101',
-  subnetMask: '255.255.255.231',
-  macAddress: 'US:B0:00:00:00:00',
-  type: Networking.INTERFACE_ETHERNET,
-}
-
 const mockWifiList = [
   { ...Fixtures.mockWifiNetwork, ssid: 'foo', active: true },
   { ...Fixtures.mockWifiNetwork, ssid: 'bar', active: false },
@@ -64,7 +57,7 @@ describe('ConnectedNetworkInfo', () => {
   beforeEach(() => {
     mockGetNetworkInterfaces.mockReturnValue({
       wifi: initialMockWifi,
-      ethernet: initialMockEthernet,
+      ethernet: null,
     })
     mockGetWifiList.mockReturnValue(mockWifiList)
   })
@@ -74,19 +67,23 @@ describe('ConnectedNetworkInfo', () => {
   })
 
   it('should render title and description', () => {
-    const [{ getByText }] = render('/connected-network-info/mockWifi')
+    const [{ getByText }] = render(
+      '/network-setup/wifi/connected-network-info/mockWifi'
+    )
     getByText('Set up your robot')
     getByText('mockWifi')
     getByText('IP Address: 127.0.0.100')
     getByText('Subnet Mask: 255.255.255.230')
-    getByText('Mac Address: WI:FI:00:00:00:00')
+    getByText('MAC Address: WI:FI:00:00:00:00')
     getByText('Change network')
   })
 
   it('when clicking Change network button, should call mock function', () => {
-    const [{ getByRole }] = render('/connected-network-info/mockWifi')
+    const [{ getByRole }] = render(
+      '/network-setup/wifi/connected-network-info/mockWifi'
+    )
     const button = getByRole('button', { name: 'Change network' })
     fireEvent.click(button)
-    expect(mockPush).toHaveBeenCalledWith('/connect-via-wifi')
+    expect(mockPush).toHaveBeenCalledWith('/network-setup/wifi')
   })
 })