+++ protocol-designer/src/feature-flags/reducers.ts
@@ -19,10 +19,6 @@ const initialFlags: Flags = {
   PRERELEASE_MODE: process.env.OT_PD_PRERELEASE_MODE === '1' || false,
   OT_PD_DISABLE_MODULE_RESTRICTIONS:
     process.env.OT_PD_DISABLE_MODULE_RESTRICTIONS === '1' || false,
-  OT_PD_ENABLE_LIQUID_COLOR_ENHANCEMENTS:
-    process.env.OT_PD_ENABLE_LIQUID_COLOR_ENHANCEMENTS === '1' || false,
-  OT_PD_ENABLE_THERMOCYCLER_GEN_2:
-    process.env.OT_PD_ENABLE_THERMOCYCLER_GEN_2 === '1' || false,
   OT_PD_ENABLE_OT_3: process.env.OT_PD_ENABLE_OT_3 === '1' || false,
 }
 // @ts-expect-error(sa, 2021-6-10): cannot use string literals as action type